Changed directory structure Added git mode to setup

+++ setup.sh
@@ -2,25 +2,17 @@
 
 
 branch=${1-master}
+mode=${2-git}
 
 nHome="$HOME/.n"
+targetDirectory="$nHome"
 masterSwitchFile="$HOME/n-bash-on-off"
 
-archiveLocation="https://github.com/naankari/n-bash/archive/$branch.zip"
-
-downloadAs="/tmp/n-bash-$branch.zip"
-extractionDir="/tmp/n-bash-$branch-zip"
-
-targetDirectory="$nHome/scripts"
-archiveContainerDirectory="n-bash-$branch/scripts"
-
-echo "Setting up nBash ..."
-
-targetDirectory="$nHome/scripts"
 if [[ -d $targetDirectory ]]; then
     echo -e "\n\nTarget directory $targetDirectory already exists. It will be backed up and deleted."
     echo "Enter 'y' or 'yes' to continue:"
     read input
+
     if [[ $input != "y" && $input != "yes" ]]; then
         echo "Exiting."
         exit 1
@@ -35,19 +27,31 @@ if [[ -d $targetDirectory ]]; then
     rm -rf "$targetDirectory"
 fi
 
-echo -e "\n\nDownloading archive from $archiveLocation as $downloadAs ..."
-rm -rf "$downloadAs"
-wget -O "$downloadAs" "$archiveLocation"
+if [[ "$mode" == "archive" ]]; then
+    archiveLocation="https://github.com/naankari/n-bash/archive/$branch.zip"
+    downloadAs="/tmp/n-bash-$branch.zip"
+    extractionDir="/tmp/n-bash-$branch-zip"
+    archiveContainerDirectory="n-bash-$branch"
 
-echo -e "\n\nExtracing archive $downloadAs in $extractionDir ..."
-rm -rf "$extractionDir"
-unzip "$downloadAs" -d "$extractionDir"
+    echo -e "\n\nDownloading archive from $archiveLocation as $downloadAs ..."
+    rm -rf "$downloadAs"
+    wget -O "$downloadAs" "$archiveLocation"
 
-echo -e "\n\nMoving contents from $extractionDir/$archiveContainerDirectory to $targetDirectory ..."
-mkdir -p "$targetDirectory"
-mv "$extractionDir/$archiveContainerDirectory"/* "$targetDirectory/"
-echo "Done copying files."
+    echo -e "\n\nExtracing archive $downloadAs in $extractionDir ..."
+    rm -rf "$extractionDir"
+    unzip "$downloadAs" -d "$extractionDir"
 
+    echo -e "\n\nMoving contents from $extractionDir/$archiveContainerDirectory to $targetDirectory ..."
+    mkdir -p "$targetDirectory"
+    mv "$extractionDir/$archiveContainerDirectory"/* "$targetDirectory/"
+    echo "Done copying files."
+else
+    git clone https://github.com/naankari/n-bash.git "$targetDirectory"
+    git checkout "$branch"
+fi
+
+
+echo "Setting up nBash ..."
 echo -e "\n\nCreating master switch file to turn on/off nBash?"
 echo "Enter 'n' or 'no' to skip:"
 read input
@@ -64,7 +68,9 @@ echo -e "[Important: Its always better to put this on top.]\n\n"
 echo "################ SETTING UP N BASH ################"
 echo "#################### FROM HERE ####################"
 echo "export N_HOME=\"$nHome\""
-echo "source \"\$N_HOME/scripts/n.sh\""
+echo "source \"\$N_HOME/n.sh\""
 echo "#################### TILL HERE ####################"
 
 echo -e "\n\nnBash setup completed."
+
+
